experiments/scripts/research/integration_implementation_guide.py
# SnapTXT 연구 기반 전처리 통합 가이드
# Integration Guide for Research-Based Preprocessing in SnapTXT

"""
easyocr_worker.py에 연구 기반 전처리를 통합하는 구체적 방법
차근차근 1개씩 단계별 구현 가이드
"""

# 1단계: 필요한 패키지 추가 설치
"""
pip install scikit-image>=0.19.0
pip install opencv-python>=4.8.0
pip install paddleocr  # 선택사항 (PaddleOCR 전처리 사용시)
"""

# 2단계: easyocr_worker.py 수정 방법

# 기존 파일에 추가할 import 구문
ADDITIONAL_IMPORTS = """
# 연구 기반 전처리 모듈 import 추가
from research_based_preprocessing_integration import KoreanOCROptimizedPreprocessor
import cv2
import numpy as np
from skimage import filters, morphology, restoration, exposure
import logging
"""

# 3단계: EasyOCRWorker 클래스 수정
EASYOCR_WORKER_MODIFICATION = """
class EasyOCRWorker:
    def __init__(self, languages=['ko', 'en'], use_gpu=False, model_storage_directory=None):
        # 기존 초기화...
        
        # 연구 기반 전처리기 추가
        self.research_preprocessor = KoreanOCROptimizedPreprocessor()
        self.use_research_preprocessing = True  # 연구 기반 전처리 활성화
        
        print("✅ Research-based preprocessing integration activated")
"""

# 4단계: 이미지 전처리 함수 통합 수정
PREPROCESSING_INTEGRATION = """
def preprocess_image_advanced(self, image, stage=None):
    \"\"\"
    통합된 전처리 시스템: 연구 기반 + 기존 4단계
    \"\"\"
    try:
        debug_info = {"stage": stage, "shape": image.shape}
        
        # 1단계: 연구 기반 전처리 적용 (새로 추가)
        if self.use_research_preprocessing:
            print(f"🔬 Applying research-based preprocessing...")
            image = self.research_preprocessor.process_image(image)
            debug_info["research_preprocessing"] = "applied"
        
        # 2단계: 기존 4단계 시스템 적용 (유지)
        if stage and stage >= 2:
            # 기존의 advanced_korean_text_processor 적용
            processed_text = self.advanced_korean_text_processor(
                text="", # OCR 후에 텍스트 처리
                original_confidence=0.5,
                stage=stage
            )
        
        debug_info["final_shape"] = image.shape
        print(f"✅ Integrated preprocessing completed: {debug_info}")
        
        return image
        
    except Exception as e:
        print(f"❌ Preprocessing integration error: {e}")
        # 실패시 기존 방법으로 폴백
        return self.preprocess_image_basic(image)
"""

# 5단계: OCR 실행 함수 수정
OCR_EXECUTION_MODIFICATION = """
def extract_text_from_image(self, image_path, **kwargs):
    \"\"\"
    연구 기반 전처리가 통합된 OCR 실행
    \"\"\"
    try:
        # 1단계: 이미지 로드
        if isinstance(image_path, str):
            image = cv2.imread(image_path)
        else:
            image = image_path
        
        if image is None:
            return self._create_error_result("이미지 로드 실패")
        
        # 2단계: 통합 전처리 적용
        print("🔄 Starting integrated preprocessing (Research-based + 4-stage)...")
        preprocessed_image = self.preprocess_image_advanced(image, stage=4)
        
        # 3단계: EasyOCR 실행
        print("🔍 Running EasyOCR with preprocessed image...")
        results = self.reader.readtext(
            preprocessed_image,
            detail=1,
            paragraph=True,
            width_ths=0.9,
            height_ths=0.9,
            **kwargs
        )
        
        # 4단계: 기존 후처리 적용
        processed_results = []
        total_confidence = 0
        
        for idx, (bbox, text, confidence) in enumerate(results):
            if confidence > 0.3:  # 신뢰도 임계값
                # 기존 한국어 후처리 적용
                enhanced_text = self.advanced_korean_text_processor(
                    text=text,
                    original_confidence=confidence,
                    stage=4
                )
                
                processed_results.append({
                    'text': enhanced_text,
                    'confidence': confidence,
                    'bbox': bbox,
                    'preprocessing': 'research_based + 4_stage'
                })
                total_confidence += confidence
        
        # 5단계: 최종 결과 반환
        final_confidence = total_confidence / len(processed_results) if processed_results else 0
        combined_text = ' '.join([result['text'] for result in processed_results])
        
        return {
            'text': combined_text,
            'confidence': final_confidence,
            'details': processed_results,
            'preprocessing_method': 'research_based_integration',
            'performance_score': min(95, final_confidence * 100)  # 최대 95점 (GPT5.2 수준)
        }
        
    except Exception as e:
        print(f"❌ OCR execution error: {e}")
        return self._create_error_result(str(e))
"""

# 6단계: 성능 평가 및 비교 함수
PERFORMANCE_EVALUATION = """
def compare_preprocessing_methods(self, image_path, test_text=None):
    \"\"\"
    연구 기반 vs 기존 방법 성능 비교
    \"\"\"
    results = {}
    
    try:
        image = cv2.imread(image_path)
        
        # 1. 기존 4단계만 사용
        self.use_research_preprocessing = False
        result_original = self.extract_text_from_image(image)
        results['original_4_stage'] = {
            'text': result_original['text'],
            'confidence': result_original['confidence'],
            'score': result_original.get('performance_score', 0)
        }
        
        # 2. 연구 기반 + 4단계 통합
        self.use_research_preprocessing = True  
        result_integrated = self.extract_text_from_image(image)
        results['research_based_integrated'] = {
            'text': result_integrated['text'],
            'confidence': result_integrated['confidence'], 
            'score': result_integrated.get('performance_score', 0)
        }
        
        # 3. 성능 향상 계산
        original_score = results['original_4_stage']['score']
        integrated_score = results['research_based_integrated']['score']
        improvement = integrated_score - original_score
        
        results['performance_comparison'] = {
            'original_score': original_score,
            'integrated_score': integrated_score,
            'improvement': improvement,
            'target_achieved': integrated_score >= 95  # GPT5.2 수준 달성 여부
        }
        
        # 4. 결과 출력
        print("=" * 60)
        print("🔬 PREPROCESSING METHOD COMPARISON RESULTS")
        print("=" * 60)
        print(f"📊 Original 4-stage:     {original_score:.1f}/100")
        print(f"🎯 Research-based:       {integrated_score:.1f}/100") 
        print(f"📈 Improvement:          +{improvement:.1f} points")
        print(f"🏆 GPT5.2 Target (95):   {'✅ ACHIEVED' if integrated_score >= 95 else '❌ NOT YET'}")
        print("=" * 60)
        
        if test_text:
            print(f"📝 Expected: {test_text}")
            print(f"🔤 Original: {results['original_4_stage']['text']}")
            print(f"🎯 Research: {results['research_based_integrated']['text']}")
        
        return results
        
    except Exception as e:
        print(f"❌ Performance comparison error: {e}")
        return {"error": str(e)}
"""

# 7단계: 점진적 구현 가이드
IMPLEMENTATION_STEPS = """
차근차근 1개씩 구현 순서:

1️⃣ 패키지 설치
   pip install scikit-image opencv-python

2️⃣ research_based_preprocessing_integration.py 파일 생성 완료

3️⃣ easyocr_worker.py에 import 추가
   from research_based_preprocessing_integration import KoreanOCROptimizedPreprocessor

4️⃣ __init__ 함수에 전처리기 초기화 추가
   self.research_preprocessor = KoreanOCROptimizedPreprocessor()

5️⃣ preprocess_image_advanced 함수 수정
   연구 기반 전처리 + 기존 4단계 통합

6️⃣ extract_text_from_image 함수 수정  
   통합 전처리 적용

7️⃣ 성능 테스트 및 비교
   compare_preprocessing_methods 함수 사용

8️⃣ 필요시 매개변수 튜닝
   한국어 특성에 맞게 세부 조정
"""

print("📋 SnapTXT Research-Based Preprocessing Integration Guide")
print("=" * 60)
print("🎯 목표: 92/100 → 95/100 (GPT5.2 수준)")
print("🔬 방법: 연구 검증된 OpenCV + scikit-image + 기존 4단계")
print("📈 예상 향상: +3점 (연구 기반 알고리즘으로)")
print("=" * 60)
print("\n다음 단계: easyocr_worker.py 파일 수정 시작")
print("차근차근 1개씩 진행하여 통합 구현 완료!")